finish codegen, edit workflows, new utils, etc. (work for 05/23/2025)

+++ src/parser/ast.rs
@@ -44,5 +44,5 @@ pub struct Identifier {
 #[derive(Debug, Clone)]
 pub struct NumericLiteral {
     pub kind: NodeType, // Always NodeType::NumericLiteral
-    pub value: f64,
+    pub value: i64,
 }
